refactor(MP-79): Remove unnecessary react imports. Include order state enum in files.
