+++ src/components/atoms/OrderDocumentTile/index.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { Box, Flex, Image, Text, Tooltip } from '@chakra-ui/react';
 import { isSummaryDocument, OrderDocument } from 'model/order/order-types';
